@@ -1,12 +1,13 @@
-import React from 'react';
 import Icon from '@economist/component-icon';
+import React from 'react';
 /* eslint-disable id-match */
 /* eslint-disable  react/display-name */
 import { createI13nNode } from 'react-i13n';
 
-const defaultPreventer = (ev) => {
-  ev.preventDefault();
-};
+function defaultPreventer(event) {
+  event.preventDefault();
+}
+
 export default function Button(props) {
   const { className, children, disabled, shadow, icon, unstyled, i13nModel } = props;
   const extraClassNames = className ? className.split(/\s+/g) : [];
@@ -22,18 +23,17 @@ export default function Button(props) {
     extraClassNames.push('link-button--shadow');
   }
   let content = children;
-  const linkProps = {...props};
-
+  let linkProps = { ...props };
   if (icon) {
     extraClassNames.push('link-button--icon');
-    extraClassNames.push(`link-button-icon--${icon.icon}`);
+    extraClassNames.push(`link-button-icon--${ icon.icon }`);
     if (icon.className) {
       extraClassNames.push(icon.className);
     }
     if (icon.color) {
-      extraClassNames.push(`icon--${icon.icon}-${icon.color}`);
+      extraClassNames.push(`icon--${ icon.icon }-${ icon.color }`);
     }
-    if (!icon.useBackground) {
+    if (Boolean(icon.useBackground) === false) {
       content = (
         <span className="link-button__group">
           <Icon {...icon} key="link-button__icon" />
@@ -44,32 +44,30 @@ export default function Button(props) {
       extraClassNames.push('link-button--icon-background');
       linkProps.style = { backgroundRepeat: 'no-repeat' };
     }
-    // We don't want this prop spreaded on <a> tag.
-    delete linkProps.icon;
+    // We don't want the icon prop to spread on <a> tag.
+    linkProps = {
+      className,
+      children,
+      disabled,
+      shadow,
+      unstyled,
+      i13nModel,
+    };
   }
 
-  linkProps.role = "button";
+  linkProps.role = 'button';
   linkProps.onClick = onClick;
-  linkProps.className= [ 'link-button' ].concat(extraClassNames).join(' ');
+  linkProps.className = [ 'link-button' ].concat(extraClassNames).join(' ');
 
-  if (props.i13nModel) {
+  if (i13nModel) {
     const I13nLink = createI13nNode('a', {
       isLeafNode: true,
       bindClickEvent: true,
       follow: true,
     });
-    return (
-      <I13nLink {...linkProps}>
-        {content}
-      </I13nLink>
-    );
-  } else {
-    return (
-      <a {...linkProps}>
-      {content}
-      </a>
-    );
+    return (<I13nLink {...linkProps}>{content}</I13nLink>);
   }
+  return (<a {...linkProps}>{content}</a>);
 }
 
 Button.propTypes = {
@@ -80,6 +78,7 @@ Button.propTypes = {
   disabled: React.PropTypes.bool,
   shadow: React.PropTypes.bool,
   unstyled: React.PropTypes.bool,
-  icon: React.PropTypes.object,
-  i13nModel: React.PropTypes.object,
+  icon: React.PropTypes.shape(Icon.propTypes),
+  // i13n genuinely takes any object
+  i13nModel: React.PropTypes.object, // eslint-disable-line react/forbid-prop-types
 };
