@@ -1,5 +1,21 @@
-module.exports = function(config) {
-  const browsers = {
+const packageJson = require('./package.json');
+const path = require('path');
+const browserifyIstanbul = require('browserify-istanbul');
+const isparta = require('isparta');
+const oneSecondInMilliseconds = 1000;
+const oneMinuteInSeconds = 60;
+const twoMinutesInMilliseconds = oneSecondInMilliseconds * oneMinuteInSeconds * 2;
+function configureBuildValue() {
+  if (process.env.GO_PIPELINE_NAME && process.env.GO_PIPELINE_LABEL) {
+    return `${ process.env.GO_PIPELINE_NAME }-${ process.env.GO_PIPELINE_LABEL }`;
+  }
+  return `localbuild-${ new Date().toJSON() }`;
+}
+module.exports = function configureKarma(config) {
+  const localBrowsers = [
+    'PhantomJS',
+  ];
+  const sauceLabsBrowsers = {
     SauceChromeLatest: {
       base: 'SauceLabs',
       browserName: 'Chrome',
@@ -20,10 +36,6 @@ module.exports = function(config) {
       base: 'SauceLabs',
       browserName: 'MicrosoftEdge',
     },
-    SauceIphoneLatest: {
-      base: 'SauceLabs',
-      browserName: 'iPad',
-    },
     SauceAndroidLatest: {
       base: 'SauceLabs',
       browserName: 'Android',
@@ -31,35 +43,57 @@ module.exports = function(config) {
   };
   config.set({
     basePath: '',
-    frameworks: ['mocha', 'chai'],
+    browsers: localBrowsers,
+    logLevel: config.LOG_INFO,
+    frameworks: [ 'browserify', 'mocha' ],
     files: [
-      require.resolve('chai-spies/chai-spies'),
-      require.resolve('chai-things/lib/chai-things'),
-      'testbundle.js'
+      path.join(packageJson.directories.test, '*.js'),
     ],
-    exclude: [
-    ],
+    exclude: [],
     preprocessors: {
+      [path.join(packageJson.directories.test, '*.js')]: [ 'browserify' ],
     },
-    reporters: ['progress', 'saucelabs'],
+    reporters: [ 'mocha', 'coverage' ],
     port: 9876,
     colors: true,
-    logLevel: config.LOG_INFO,
+    concurrency: 3,
     autoWatch: false,
-    customLaunchers: browsers,
-    browserDisconnectTimeout: 1000 * 60 * 2,
-    browserNoActivityTimeout: 1000 * 60 * 2,
-    sauceLabs: {
-      testName: require('./package').name,
-      startConnect: true,
-      build: (function () {
-        if (process.env.GO_PIPELINE_NAME && process.env.GO_PIPELINE_LABEL) {
-          return process.env.GO_PIPELINE_NAME + '-' + process.env.GO_PIPELINE_LABEL;
-        }
-        return 'localbuild-' + new Date().toJSON();
-      })(),
+    captureTimeout: twoMinutesInMilliseconds,
+    browserDisconnectTimeout: twoMinutesInMilliseconds,
+    browserNoActivityTimeout: twoMinutesInMilliseconds,
+    singleRun: true,
+    browserify: {
+      transform: [
+        browserifyIstanbul({
+          instrumenter: isparta,
+          ignore: [ '**/node_modules/**', '**/test/**' ],
+        }),
+        'babelify',
+      ],
+      configure: (bundle) => {
+        bundle.on('prebundle', () => {
+          bundle.external('react/lib/ReactContext');
+          bundle.external('react/lib/ExecutionEnvironment');
+        });
+      },
     },
-    browsers: Object.keys(browsers),
-    singleRun: true
-  })
-}
+    coverageReporter: {
+      type: 'lcov',
+      dir: 'coverage',
+    },
+  });
+
+  if (process.env.SAUCE_ACCESS_KEY && process.env.SAUCE_USERNAME) {
+    config.reporters.push('saucelabs');
+    config.set({
+      customLaunchers: sauceLabsBrowsers,
+      browsers: Object.keys(sauceLabsBrowsers),
+      sauceLabs: {
+        testName: packageJson.name,
+        recordVideo: true,
+        startConnect: true,
+        build: configureBuildValue(),
+      },
+    });
+  }
+};
